TASK: Provide tests that subtree tag moving does not create obsolete edges
